# web/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Laplace.Endpoints.OpenAICompat | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/chess/legal": {
      "post": {
        "tags": [
          "chess"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/move": {
      "post": {
        "tags": [
          "chess"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/eval": {
      "post": {
        "tags": [
          "chess"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/bestmove": {
      "post": {
        "tags": [
          "chess"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BestMoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/explore": {
      "post": {
        "tags": [
          "chess"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExploreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/train/start": {
      "post": {
        "tags": [
          "chess"
        ],
        "parameters": [
          {
            "name": "temperature",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "weight",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "maxPlies",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "games",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/chess/play/start": {
      "post": {
        "tags": [
          "chess"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/play/move": {
      "post": {
        "tags": [
          "chess"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayMoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/play/bestmove": {
      "post": {
        "tags": [
          "chess"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayBestMoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/play/finish": {
      "post": {
        "tags": [
          "chess"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/lichess/games/{gameId}/chat": {
      "get": {
        "tags": [
          "chess"
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/chess/lichess/start": {
      "post": {
        "tags": [
          "chess"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LichessStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/chess/lab/start": {
      "post": {
        "tags": [
          "chess-lab"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/lab/stop/{jobId}": {
      "post": {
        "tags": [
          "chess-lab"
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/chess/lab/jobs/{jobId}": {
      "get": {
        "tags": [
          "chess-lab"
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/lab/jobs/{jobId}/events": {
      "get": {
        "tags": [
          "chess-lab"
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/lab/jobs/{jobId}/artifact/{name}": {
      "get": {
        "tags": [
          "chess-lab"
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/chess/lab/jobs/{jobId}/ingest": {
      "post": {
        "tags": [
          "chess-lab"
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BestMoveRequest": {
        "required": [
          "fen",
          "temperature",
          "depth",
          "substrate",
          "moves"
        ],
        "type": "object",
        "properties": {
          "fen": {
            "type": "string"
          },
          "temperature": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "depth": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "substrate": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "moves": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "EvalRequest": {
        "required": [
          "fen",
          "depth",
          "substrate"
        ],
        "type": "object",
        "properties": {
          "fen": {
            "type": "string"
          },
          "depth": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "substrate": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
      "ExploreRequest": {
        "required": [
          "fen",
          "player",
          "limit"
        ],
        "type": "object",
        "properties": {
          "fen": {
            "type": "string"
          },
          "player": {
            "type": [
              "null",
              "string"
            ]
          },
          "limit": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "FenRequest": {
        "required": [
          "fen"
        ],
        "type": "object",
        "properties": {
          "fen": {
            "type": "string"
          }
        }
      },
...
      "LabStartRequest": {
        "required": [
          "kind",
          "config"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": [
              "null",
              "string"
            ]
          },
          "config": {
            "type": [
              "null",
              "object"
            ]
          }
        }
      },
...
      "LichessStartRequest": {
        "required": [
          "depth",
          "maxConcurrent",
          "substrate",
          "speeds"
        ],
        "type": "object",
        "properties": {
          "depth": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "maxConcurrent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "substrate": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "speeds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "MoveRequest": {
        "required": [
          "fen",
          "uci",
          "moves"
        ],
        "type": "object",
        "properties": {
          "fen": {
            "type": "string"
          },
          "uci": {
            "type": "string"
          },
          "moves": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PlayBestMoveRequest": {
        "required": [
          "sessionId",
          "fen",
          "depth",
          "substrate"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "fen": {
            "type": "string"
          },
          "depth": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "substrate": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
      "PlayFinishRequest": {
        "required": [
          "sessionId",
          "status",
          "adjudicated"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": [
              "null",
              "string"
            ]
          },
          "adjudicated": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
      "PlayMoveRequest": {
        "required": [
          "sessionId",
          "fen",
          "uci"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "fen": {
            "type": "string"
          },
          "uci": {
            "type": "string"
          }
        }
      },
      "PlayStartRequest": {
        "required": [
          "record",
          "moves"
        ],
        "type": "object",
        "properties": {
          "record": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "moves": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}